alterações pedidas
