+++ exceptions/customer_exceptions.py
@@ -1,0 +1,4 @@
+class EntityNotFound(Exception):
+    def __init__(self, message: str):
+        self.message = message
+        super().__init__(self.message)
